Centralize CLI commands in one single `ota` command
