+++ bin/ota-track.js
@@ -1,7 +1,6 @@
 #! /usr/bin/env node
-import './.env.js'; // Workaround to ensure `SUPPRESS_NO_CONFIG_WARNING` is set before config is imported
+import './.env.js';
 
-import fs from 'fs';
 import path from 'path';
 import { fileURLToPath, pathToFileURL } from 'url';
 
@@ -9,17 +8,13 @@ import { program } from 'commander';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
-const { description, version } = JSON.parse(fs.readFileSync(new URL('../package.json', import.meta.url)).toString());
+const track = (await import(pathToFileURL(path.resolve(__dirname, '../src/index.js')))).default; // asynchronous loading to ensure .env.js is load before
 
 program
-  .name('ota-track')
-  .description(description)
-  .version(version)
+  .description('Retrieve declared documents, then record snapshots and extract versions, then publish the resulting records')
   .option('-s, --services [serviceId...]', 'service IDs of services to handle')
   .option('-d, --documentTypes [documentType...]', 'terms types to handle')
   .option('-r, --refilter-only', 'only refilter exisiting snapshots with last declarations and engine\'s updates')
   .option('--schedule', 'schedule automatic document tracking');
-
-const track = (await import(pathToFileURL(path.resolve(__dirname, '../src/index.js')))).default;
 
 track(program.parse(process.argv).opts());
